+++ test/view-spec.js
@@ -192,3 +192,30 @@ it("Falls back to center when no data and zoomToData given", async () => {
 	await expect(center.lat).toBe(latLng.lat);
 	await expect(center.lng).toBe(latLng.lng);
 });
+
+it("Falls back to center when no zoomToData given", async () => {
+	const data = {type: "FeatureCollection", features: []};
+	const latLng = {lat: 25, lng: 60};
+	const options = {
+		center: latLng,
+		data: {geoData: {type: "Point", coordinates: [45, 70]}}
+	};
+
+	const map = await createMap({
+		...options
+	});
+
+	const center = await map.e("map.getCenter()");
+	await expect(center.lat).toBe(latLng.lat);
+	await expect(center.lng).toBe(latLng.lng);
+});
+
+it("keeps finnish tileLayer if center is outside Finland but zoomToData causes view to initialize into Finland", async () => {
+	const congo =  {
+		"lat": 79.3499057749654,
+		"lng": 21.160612106323246
+	};
+	const data = {geoData: {type: "Point", coordinates: [25, 60]}};
+	const map = await createMap({tileLayerName: "taustakartta", zoom: 4, center: congo, data, zoomToData: true});
+	await expect(await map.e("tileLayerName")).toBe("taustakartta");
+});